tool_use/ml_agent/requirements.txt
# ML Agent Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# LLM Integration
aisuite>=0.1.0
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Optional: For real TCGA data download
# gdc-client==1.6.1
